api/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29
pydantic>=2.6
mysql-connector-python>=8.1
